+++ packages/sources/react/src/components/overlays/VtmnSnackbar/index.ts
@@ -1,0 +1,2 @@
+export { default as VtmnSnackbar } from './VtmnSnackbar';
+export * from './VtmnSnackbar';